design: startpage color change, login backbutton add
